first working but non-portable attempt to fetch from CPAN
We need a lot of additional depenecies here, like HTTP::UA to fetch dists
when the server gives us chunks or like Archive::Tar and Compress::Zlib for
unpacking the dists itself.
These dists are problematic right now when it comes to portability. A real
chance that it works is only given on 64bit linux atm.