implement Panda::Reporter, which saves reports to disk for now
This is WIP. It will create a test report file once a dist installation is attempted.
See this example of a FAIL: https://gist.github.com/FROGGS/5b25776e1ce3f297cb05
Next step is to submit these reports for review to a web service.
We also need to ask the user for permission to generate and send these reports.