Remove action-less shortcut for 'install'
Having panda operate on a bunch of actions that aren't prefixed with
special characters (e.g. --smoke or --install) nor otherwise
distinguished from valid package names makes the use of 'panda Foo::Bar'
as a shortcut for 'panda install Foo::Bar' is really not a good idea.
For example, 'panda help' used to be interpreted as trying to install a
'help' package. So it's gone, requiring an explicit 'install' to be
given to panda now.