No longer copy anything to blib when building
No one seems to know why we even copy files to a "blib" directory during build.
By installing from the directory as-is we get rid of the difference between a
local build and the build done by panda which most importantly means that
the relative path from libraries to the resources directory is the same in both
cases. This greatly simplifies makefiles of modules needing to generate files
during the build process.