Supply relative paths to CompUnit::Repository::Installation::install
Precompiled modules are checked against their source files on loading to detect
outdated precomp files. This is done by recording the source file's id. Thus
this id must be stable and survive reinstallation of a module. Otherwise the
precompilation file's modification time will be compared to an outdated source
file.

This also means that the source file's path must be stable. Therefore we have
to supply relative paths to the install method, as the absolute paths contain
panda's work directory which contains a time stamp.